chore: break routes into nested routes
create controllers for each sub level of the organization
