+++ avoda-be/src/organizations/organizations.module.ts
@@ -4,13 +4,15 @@ import { DatabaseModule } from '@/database/database.module';
 
 import { UserService } from '@/users/users.service';
 import { userProviders } from '@/users/users.provider';
+import { organizationProvider } from './organizations.provider';
 
 import { OrganizationsService } from './services/organizations.service';
 import { PermissionsService } from './services/permissions.service';
 import { MemberService } from './services/members.service';
 
-import { OrganizationsController } from './organizations.controller';
-import { organizationProvider } from './organizations.provider';
+import { OrganizationsController } from './controllers/organizations.controller';
+import { PermissionsController } from './controllers/permissions.controller';
+import { MembersController } from './controllers/members.controller';
 
 @Module({
   imports: [DatabaseModule],
@@ -22,6 +24,10 @@ import { organizationProvider } from './organizations.provider';
     ...organizationProvider,
     ...userProviders,
   ],
-  controllers: [OrganizationsController],
+  controllers: [
+    OrganizationsController,
+    PermissionsController,
+    MembersController,
+  ],
 })
 export class OrganizationsModule {}
